Xinder :: Re-Worked, for December.
Xinder has been re-worked for a December, (Christmas-y Gift) and a Parting Gift. Therefore, here is that. Xinder will be the last confirmed project under the "L-XDT" brand, because to be frank, I have exhausted all the social medias I could think off and I need to focus on other aspects on my life such as learning Java. Thanks for being part of the Journey!

- Alvin Maranx, II ツ (Avun), Owner of Leidenschaft, Xiotim & L-XDT Respectively.