The End of L-XDT is Nigh: The Added Avunii Button.
This is the end of an era for L-XDT. Showing that the end is neigh, we need to look for new resorts, to show that there is still work in progress. For this I added the Avunii button, where you can find my work after you are done checking out the L-XDT catalogue. Give me a minute, you may be shocked.